New API! More abstraction, less boilerplate code.
Zoomit.getContentInfo() is no longer just a thinly-veiled JSONP helper.

It now encapsulates the parsing of the non-RESTful response object, e.g.
determining whether there was an error or not, as well as the boilerplate
processing of the content object, e.g. determining whether it's ready or not.

Its method signature has thus changed from (url, callback) to just (opts),
where the options are "id" or "url" as input (one of these is required) and
"error", "ready", "failed" and "processing" as callbacks.

Also updated the sample page and brought the license copyright to 2011.